refactored vibes frontend, created vibeUtil to clean up vibe component
